[NF] +/-/>/< commands compaction was added with an external test launcher
