+++ .gitmodules
@@ -1,0 +1,3 @@
+[submodule "test/external_data/Brainfuck"]
+	path = test/external_data/Brainfuck
+	url = https://github.com/rdebath/Brainfuck
