--- .idea/.gitignore
@@ -1,10 +1,0 @@
-# Default ignored files
-/shelf/
-/workspace.xml
-# Datasource local storage ignored files
-/dataSources/
-/dataSources.local.xml
-# Editor-based HTTP Client requests
-/httpRequests/
-# Zeppelin ignored files
-/ZeppelinRemoteNotebooks/
