+++ README.md
@@ -2,9 +2,12 @@
 Translator of Brainfuck to Javascript
 
 # How to run
-set GOPATH=.
-go run src/main/main.go test/data/HelloWorld.bf test/data/HelloWorld.in
-node tmp/translated.js
+    set GOPATH=.
+    go run src/main/main.go test/data/ReverseInput.bf test/data/ReverseInput.in
+    node tmp/translated.js
 
 # How to test
-go test -v ./...
+    go test -v .\test\translator-service-test\...
+
+# How to test with external repo
+    git submodule add https://github.com/rdebath/Brainfuck .\test\external_data\
